backend/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
mysqlclient==2.2.4
django-cors-headers==4.3.1
firebase-admin==6.4.0